Cleanup, find meaningful TODOs.
